examples/hello-world-local/pyproject.toml
[project]
name = "hello-world-local"
version = "0.1.0"
description = "Hello World - Local agent execution without server deployment"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "glaip-sdk[local,privacy,memory]",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
